third_party/WebKit/LayoutTests/bluetooth/service/getCharacteristics/gen-garbage-collection-ran-during-error.html

<!-- Generated by //third_party/WebKit/LayoutTests/bluetooth/generate.py -->
<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
let test_desc = 'Garbage Collection ran during getCharacteristics call that ' +
   'fails. Should not crash';
let func_promise;
bluetooth_test(() => getHealthThermometerService()
  .then(({service}) => {
    func_promise = assert_promise_rejects_with_message(
        service.getCharacteristics(),
        new DOMException(
            'GATT Server is disconnected. Cannot retrieve characteristics. ' +
            '(Re)connect first with `device.gatt.connect`.',
            'NetworkError'));
    // Disconnect called to clear attributeInstanceMap and allow the object to
    // get garbage collected.
    service.device.gatt.disconnect();
  })
  .then(runGarbageCollection)
  .then(() => func_promise), test_desc);

</script>
